Step 2: Draw a Seagl
In this step we add our first Component: Seagl.

Then we added two entities to our game: Camera and Seagl.

Both entities used the Transform state, which required us to add the
Transform bundle to the game.

We also loaded a sprite sheet and attached our first Sprite to our Seagl entity.