Step 4: Put a burger on it
Now we go through the previous steps of creating a component, entity,
and system for a new feature: eating burgers!

The only real difference between this system and the last one is we're
seeing how to iterate over two sets of entities, and how to delete
entities in a system.